+++ CHANGELOG.md
@@ -5,6 +5,7 @@ x.y.z Release notes (yyyy-MM-dd)
 * Mac Catalyst and arm64 simulators are now supported when integrating with Cocoapods. 
   The mininum required version of Cocoapods supported is 1.10
 * Add Xcode 12.1 binaries to the release package.
+* Add Combine support for `Realm.asyncOpen()`.
 
 ### Fixed
 * Implement precise and unbatched notification of sync completion events. This avoids a race condition where an earlier upload completion event will notify a later waiter whose changes haven't been uploaded yet. ([#1118](https://github.com/realm/realm-object-store/pull/1118)).